requirements.txt
mysqlclient
python-dotenv
mysql-connector-python